abstract lock service tests
to allow for testing other lock service impls
